don't mess with sites that have no network set

+++ util/merge_coop_nwsli.py
@@ -99,11 +99,12 @@ for row in hcursor:
           network, gtxt))
     mcursor.close()
     MESOSITE.commit()
-    hcursor2.execute("""DELETE from unknown where nwsli = '%s'""" % (nwsli,))
+    if network != "":
+        hcursor2.execute("""DELETE from unknown where nwsli = '%s'""" % (nwsli,))
     
-    cmd = "/usr/bin/env python /var/www/scripts/util/addSiteMesosite.py %s %s" % (network, nwsli)
-    os.system(cmd)
-    print 'Added %s %s [%s]' % (nwsli, network, sites[nwsli]['name'])
+        cmd = "/usr/bin/env python /var/www/scripts/util/addSiteMesosite.py %s %s" % (network, nwsli)
+        os.system(cmd)
+        print 'Added %s %s [%s]' % (nwsli, network, sites[nwsli]['name'])
     
 MESOSITE.commit()
 HADS.commit()
